+++ crates/holochain/src/core/workflow/app_validation_workflow/unit_tests.rs
@@ -13,13 +13,12 @@ use crate::{
     sweettest::{SweetDnaFile, SweetInlineZomes},
 };
 use fixt::fixt;
-use holo_hash::{AgentPubKey, HashableContentExtSync};
+use holo_hash::{AgentPubKey, AnyDhtHash, HashableContentExtSync};
 use holochain_p2p::{HolochainP2pDnaFixturator, MockHolochainP2pDnaT};
 use holochain_state::{host_fn_workspace::HostFnWorkspaceRead, mutations::insert_op};
 use holochain_types::{
     chain::MustGetAgentActivityResponse,
     dht_op::{DhtOp, DhtOpHashed, WireOps},
-    dna::DnaFile,
     record::WireRecordOps,
 };
 use holochain_wasmer_host::module::ModuleCache;
@@ -80,7 +79,7 @@ async fn validation_callback_must_get_action() {
     } = TestCase::new(zomes).await;
 
     let network = Arc::new(fixt!(HolochainP2pDna));
-    let fetched_dependencies = Arc::new(Mutex::new(ValidationDependencies::new()));
+    let validation_dependencies = Arc::new(Mutex::new(ValidationDependencies::new()));
 
     // a create by alice
     let mut create = fixt!(Create);
@@ -91,6 +90,11 @@ async fn validation_callback_must_get_action() {
     delete.author = bob.clone();
     delete.deletes_address = create_action.clone().to_hash();
     let delete_action_signed_hashed = SignedHashed::new_unchecked(delete.clone(), fixt!(Signature));
+    let delete_dht_op = DhtOp::RegisterDeletedBy(
+        delete_action_signed_hashed.signature.clone(),
+        delete.clone(),
+    );
+    let delete_dht_op_hash = delete_dht_op.to_hash();
     let delete_action_op = Op::RegisterDelete(RegisterDelete {
         delete: delete_action_signed_hashed.clone(),
         original_action: EntryCreationAction::Create(create.clone()),
@@ -102,10 +106,11 @@ async fn validation_callback_must_get_action() {
     // validation should indicate it is awaiting create action hash
     let outcome = run_validation_callback(
         invocation.clone(),
+        &delete_dht_op_hash,
         &ribosome,
         workspace.clone(),
         network.clone(),
-        fetched_dependencies.clone(),
+        validation_dependencies.clone(),
     )
     .await
     .unwrap();
@@ -121,10 +126,11 @@ async fn validation_callback_must_get_action() {
     // the same validation should now successfully validate the op
     let outcome = run_validation_callback(
         invocation,
+        &delete_dht_op_hash,
         &ribosome,
         workspace,
         network,
-        fetched_dependencies.clone(),
+        validation_dependencies.clone(),
     )
     .await
     .unwrap();
@@ -178,6 +184,11 @@ async fn validation_callback_awaiting_deps_hashes() {
     delete.author = bob.clone();
     delete.deletes_address = create_action.clone().to_hash();
     let delete_action_signed_hashed = SignedHashed::new_unchecked(delete.clone(), fixt!(Signature));
+    let delete_dht_op = DhtOp::RegisterDeletedBy(
+        delete_action_signed_hashed.signature.clone(),
+        delete.clone(),
+    );
+    let delete_dht_op_hash = delete_dht_op.to_hash();
     let delete_action_op = Op::RegisterDelete(RegisterDelete {
         delete: delete_action_signed_hashed.clone(),
         original_action: EntryCreationAction::Create(create.clone()),
@@ -201,16 +212,16 @@ async fn validation_callback_awaiting_deps_hashes() {
     });
 
     let network = Arc::new(network);
-
-    let fetched_dependencies = Arc::new(Mutex::new(ValidationDependencies::new()));
+    let validation_dependencies = Arc::new(Mutex::new(ValidationDependencies::new()));
 
     // app validation should indicate missing action is being awaited
     let outcome = run_validation_callback(
         invocation.clone(),
+        &delete_dht_op_hash,
         &ribosome,
         workspace.clone(),
         network.clone(),
-        fetched_dependencies.clone(),
+        validation_dependencies.clone(),
     )
     .await
     .unwrap();
@@ -223,10 +234,11 @@ async fn validation_callback_awaiting_deps_hashes() {
     // has been fetched
     let outcome = run_validation_callback(
         invocation,
+        &delete_dht_op_hash,
         &ribosome,
         workspace,
         network,
-        fetched_dependencies.clone(),
+        validation_dependencies.clone(),
     )
     .await
     .unwrap();
@@ -300,6 +312,11 @@ async fn validation_callback_awaiting_deps_agent_activity() {
     let delete_action = Action::Delete(delete.clone());
     let delete_action_signed_hashed =
         SignedActionHashed::new_unchecked(delete_action.clone(), fixt!(Signature));
+    let delete_dht_op = DhtOp::RegisterDeletedBy(
+        delete_action_signed_hashed.signature.clone(),
+        delete.clone(),
+    );
+    let delete_dht_op_hash = delete_dht_op.to_hash();
     let delete_action_op = Op::RegisterDelete(RegisterDelete {
         delete: SignedHashed::new_unchecked(delete.clone(), fixt!(Signature)),
         original_action: EntryCreationAction::Create(create.clone()),
@@ -338,15 +355,16 @@ async fn validation_callback_awaiting_deps_agent_activity() {
     });
     let network = Arc::new(network);
 
-    let fetched_dependencies = Arc::new(Mutex::new(ValidationDependencies::new()));
+    let validation_dependencies = Arc::new(Mutex::new(ValidationDependencies::new()));
 
     // app validation should indicate missing action is being awaited
     let outcome = run_validation_callback(
         invocation.clone(),
+        &delete_dht_op_hash,
         &ribosome,
         workspace.clone(),
         network.clone(),
-        fetched_dependencies.clone(),
+        validation_dependencies.clone(),
     )
     .await
     .unwrap();
@@ -358,11 +376,12 @@ async fn validation_callback_awaiting_deps_agent_activity() {
     // app validation outcome should be accepted, now that bob's missing agent
     // activity is available in alice's cache
     let outcome = run_validation_callback(
-        invocation.clone(),
+        invocation,
+        &delete_dht_op_hash,
         &ribosome,
         workspace,
         network,
-        fetched_dependencies.clone(),
+        validation_dependencies.clone(),
     )
     .await
     .unwrap();
@@ -417,6 +436,11 @@ async fn validation_callback_prevent_multiple_identical_hash_fetches() {
     delete.author = bob.clone();
     delete.deletes_address = create_action.clone().to_hash();
     let delete_action_signed_hashed = SignedHashed::new_unchecked(delete.clone(), fixt!(Signature));
+    let delete_dht_op = DhtOp::RegisterDeletedBy(
+        delete_action_signed_hashed.signature.clone(),
+        delete.clone(),
+    );
+    let delete_dht_op_hash = delete_dht_op.to_hash();
     let delete_action_op = Op::RegisterDelete(RegisterDelete {
         delete: delete_action_signed_hashed.clone(),
         original_action: EntryCreationAction::Create(create.clone()),
@@ -449,22 +473,24 @@ async fn validation_callback_prevent_multiple_identical_hash_fetches() {
     });
     let network = Arc::new(network);
 
-    let fetched_dependencies = Arc::new(Mutex::new(ValidationDependencies::new()));
+    let validation_dependencies = Arc::new(Mutex::new(ValidationDependencies::new()));
 
     // run two op validations that depend on the same record in parallel
     let validate_1 = run_validation_callback(
         invocation.clone(),
+        &delete_dht_op_hash,
         &ribosome,
         workspace.clone(),
         network.clone(),
-        fetched_dependencies.clone(),
+        validation_dependencies.clone(),
     );
     let validate_2 = run_validation_callback(
         invocation,
+        &delete_dht_op_hash,
         &ribosome,
         workspace,
         network,
-        fetched_dependencies.clone(),
+        validation_dependencies.clone(),
     );
     futures::future::join_all([validate_1, validate_2]).await;
 
@@ -473,7 +499,7 @@ async fn validation_callback_prevent_multiple_identical_hash_fetches() {
 
     assert_eq!(times_same_hash_is_fetched.load(Ordering::Relaxed), 1);
     // after successfully fetching dependencies, the set should be empty
-    assert_eq!(fetched_dependencies.lock().missing_hashes.len(), 0);
+    assert_eq!(validation_dependencies.lock().missing_hashes.len(), 0);
 }
 
 #[tokio::test(flavor = "multi_thread")]
@@ -542,6 +568,11 @@ async fn validation_callback_prevent_multiple_identical_agent_activity_fetches()
     let delete_action = Action::Delete(delete.clone());
     let delete_action_signed_hashed =
         SignedActionHashed::new_unchecked(delete_action.clone(), fixt!(Signature));
+    let delete_dht_op = DhtOp::RegisterDeletedBy(
+        delete_action_signed_hashed.signature.clone(),
+        delete.clone(),
+    );
+    let delete_dht_op_hash = delete_dht_op.to_hash();
     let delete_action_op = Op::RegisterDelete(RegisterDelete {
         delete: SignedHashed::new_unchecked(delete.clone(), fixt!(Signature)),
         original_action: EntryCreationAction::Create(create.clone()),
@@ -580,22 +611,24 @@ async fn validation_callback_prevent_multiple_identical_agent_activity_fetches()
     });
     let network = Arc::new(network);
 
-    let fetched_dependencies = Arc::new(Mutex::new(ValidationDependencies::new()));
+    let validation_dependencies = Arc::new(Mutex::new(ValidationDependencies::new()));
 
     // run two op validations that depend on the same record in parallel
     let validate_1 = run_validation_callback(
         invocation.clone(),
+        &delete_dht_op_hash,
         &ribosome,
         workspace.clone(),
         network.clone(),
-        fetched_dependencies.clone(),
+        validation_dependencies.clone(),
     );
     let validate_2 = run_validation_callback(
         invocation,
+        &delete_dht_op_hash,
         &ribosome,
         workspace,
         network,
-        fetched_dependencies.clone(),
+        validation_dependencies.clone(),
     );
     futures::future::join_all([validate_1, validate_2]).await;
 
@@ -604,11 +637,164 @@ async fn validation_callback_prevent_multiple_identical_agent_activity_fetches()
 
     assert_eq!(times_same_hash_is_fetched.load(Ordering::Relaxed), 1);
     // after successfully fetching dependencies, the set should be empty
-    assert_eq!(fetched_dependencies.lock().missing_hashes.len(), 0);
+    assert_eq!(validation_dependencies.lock().missing_hashes.len(), 0);
 }
 
 #[tokio::test(flavor = "multi_thread")]
-async fn hashes_missing_for_op_is_updated_with_unresolved_deps() {}
+async fn hashes_missing_for_op_are_updated_before_and_after_fetching_deps() {
+    holochain_trace::test_run().unwrap();
+
+    let zomes = SweetInlineZomes::new(vec![], 0).integrity_function("validate", {
+        move |api, op: Op| {
+            if let Op::RegisterDelete(RegisterDelete {
+                original_action, ..
+            }) = op
+            {
+                let result = api.must_get_action(MustGetActionInput(original_action.to_hash()));
+                if result.is_ok() {
+                    Ok(ValidateCallbackResult::Valid)
+                } else {
+                    Ok(ValidateCallbackResult::UnresolvedDependencies(
+                        UnresolvedDependencies::Hashes(vec![original_action.to_hash().into()]),
+                    ))
+                }
+            } else {
+                unreachable!()
+            }
+        }
+    });
+
+    let TestCase {
+        zomes_to_invoke,
+        ribosome,
+        alice,
+        workspace,
+        bob,
+        ..
+    } = TestCase::new(zomes).await;
+
+    // a create by alice
+    let mut create = fixt!(Create);
+    create.author = alice.clone();
+    let create_action = Action::Create(create.clone());
+    let create_action_signed_hashed =
+        SignedHashed::new_unchecked(create_action.clone(), fixt!(Signature));
+    // a delete by bob that references alice's create
+    let mut delete = fixt!(Delete);
+    delete.author = bob.clone();
+    delete.deletes_address = create_action.clone().to_hash();
+    let delete_action_signed_hashed = SignedHashed::new_unchecked(delete.clone(), fixt!(Signature));
+    let delete_dht_op = DhtOp::RegisterDeletedBy(
+        delete_action_signed_hashed.signature.clone(),
+        delete.clone(),
+    );
+    let delete_dht_op_hash = delete_dht_op.to_hash();
+    let delete_action_op = Op::RegisterDelete(RegisterDelete {
+        delete: delete_action_signed_hashed.clone(),
+        original_action: EntryCreationAction::Create(create.clone()),
+        original_entry: None,
+    });
+
+    // mock network that returns the requested create action
+    let mut network = MockHolochainP2pDnaT::new();
+    network.expect_get().returning({
+        let create_action_signed_hashed = create_action_signed_hashed.clone();
+        move |hash, _| {
+            assert_eq!(hash, create_action_signed_hashed.as_hash().clone().into());
+            Ok(vec![WireOps::Record(WireRecordOps {
+                action: Some(Judged::new(
+                    create_action_signed_hashed.clone().into(),
+                    ValidationStatus::Valid,
+                )),
+                deletes: vec![],
+                updates: vec![],
+                entry: None,
+            })])
+        }
+    });
+    let network = Arc::new(network);
+
+    let invocation = ValidateInvocation::new(zomes_to_invoke, &delete_action_op).unwrap();
+    let validation_dependencies = Arc::new(Mutex::new(ValidationDependencies::new()));
+
+    // hashes missing for delete dht op should be empty
+    assert_eq!(
+        validation_dependencies
+            .lock()
+            .hashes_missing_for_op
+            .get(&delete_dht_op_hash),
+        None
+    );
+
+    // filtering out ops with missing dependencies should not filter anything
+    let ops_to_validate = vec![delete_dht_op.clone().into_hashed()];
+    let filtered_ops_to_validate = validation_dependencies
+        .lock()
+        .filter_ops_missing_dependencies(ops_to_validate.clone());
+    assert_eq!(filtered_ops_to_validate, ops_to_validate);
+
+    let _ = run_validation_callback(
+        invocation.clone(),
+        &delete_dht_op_hash,
+        &ribosome,
+        workspace.clone(),
+        network.clone(),
+        validation_dependencies.clone(),
+    )
+    .await
+    .unwrap();
+
+    // hashes missing for delete dht op should contain the missing create hash
+    assert_eq!(
+        validation_dependencies
+            .lock()
+            .hashes_missing_for_op
+            .get(&delete_dht_op_hash)
+            .unwrap()
+            .clone()
+            .into_iter()
+            .collect::<Vec<AnyDhtHash>>(),
+        vec![create_action_signed_hashed.as_hash().clone().into()]
+    );
+
+    // filtering out ops with missing dependencies should filter out delete
+    let ops_to_validate = vec![delete_dht_op.clone().into_hashed()];
+    let filtered_ops_to_validate = validation_dependencies
+        .lock()
+        .filter_ops_missing_dependencies(ops_to_validate.clone());
+    assert_eq!(filtered_ops_to_validate, vec![]);
+
+    // await while missing record is being fetched in background task
+    tokio::time::sleep(Duration::from_millis(10)).await;
+
+    let _ = run_validation_callback(
+        invocation,
+        &delete_dht_op_hash,
+        &ribosome,
+        workspace,
+        network,
+        validation_dependencies.clone(),
+    )
+    .await
+    .unwrap();
+
+    // hashes missing for delete dht op should be empty again after create
+    // has been fetched
+    assert_eq!(
+        validation_dependencies
+            .lock()
+            .hashes_missing_for_op
+            .get(&delete_dht_op_hash),
+        None
+    );
+
+    // filtering out ops with missing dependencies should not filter anything
+    let ops_to_validate = vec![delete_dht_op.into_hashed()];
+    let filtered_ops_to_validate = validation_dependencies
+        .lock()
+        .filter_ops_missing_dependencies(ops_to_validate.clone());
+    assert_eq!(filtered_ops_to_validate, ops_to_validate);
+}
 
 // test case with alice and bob agent keys
 // test space created by alice